+++ record_web/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 1.1.1
 * fix: Pausing for too long with Chromium based browsers breaks the recording.
+* fix: Migrate to package:web missing parts & example.
 
 ## 1.1.0
 * feat: Migrated implementation to package:web.